Rejig SharedArrayBuffer help page section
It now mentions the most common case first: using a weird embedded
browser that breaks the application. Server misconfigurations are now
shunted away to a small paragraph, since it just confuses users into
thinking the site is wrong. Unofficial installations of the client are
rare and server owners don't need this page to figure out what to do
since it's already explained in the installation instructions.